backend/requirements.txt
# Core Dependencies
numpy>=1.24.0
pandas>=2.0.0
xarray>=2023.1.0

# Copernicus Marine Data
copernicusmarine>=1.0.0

# Weather Data
openmeteo-requests>=1.1.0
requests-cache>=1.1.0
retry-requests>=2.0.0

# NetCDF File Handling
netCDF4>=1.6.0

# Environment Variables
python-dotenv>=1.0.0

# API Framework (Backend için)
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

# CORS Support (Flutter için)
python-multipart>=0.0.6

# Optional: Machine Learning (balık tahmini için)
scikit-learn>=1.3.0
scipy>=1.11.0

# Optional: Data Visualization (geliştirme aşamasında)
matplotlib>=3.7.0
seaborn>=0.12.0

# Optional: Database (ileride gerekirse)
# sqlalchemy>=2.0.0
# psycopg2-binary>=2.9.0